Move searchQuestionnaireResponse function to default repository
