+++ android/quest/src/main/java/org/smartregister/fhircore/quest/ui/questionnaire/QuestionnaireViewModel.kt
@@ -571,7 +571,8 @@ constructor(
         !questionnaireConfig.resourceIdentifier.isNullOrEmpty() &&
         subjectType != null
     ) {
-      searchQuestionnaireResponse(
+      defaultRepository
+        .searchQuestionnaireResponse(
           resourceId = questionnaireConfig.resourceIdentifier!!,
           resourceType = questionnaireConfig.resourceType ?: subjectType,
           questionnaireId = questionnaire.logicalId,
@@ -1064,48 +1065,6 @@ constructor(
     }
   }
 
-  /**
-   * This function searches and returns the latest [QuestionnaireResponse] for the given
-   * [resourceId] that was extracted from the [Questionnaire] identified as [questionnaireId].
-   * Returns null if non is found.
-   */
-  suspend fun searchQuestionnaireResponse(
-    resourceId: String,
-    resourceType: ResourceType,
-    questionnaireId: String,
-    encounterId: String?,
-    questionnaireResponseStatus: String? = null,
-  ): QuestionnaireResponse? {
-    val search =
-      Search(ResourceType.QuestionnaireResponse).apply {
-        filter(
-          QuestionnaireResponse.SUBJECT,
-          { value = resourceId.asReference(resourceType).reference },
-        )
-        filter(
-          QuestionnaireResponse.QUESTIONNAIRE,
-          { value = questionnaireId.asReference(ResourceType.Questionnaire).reference },
-        )
-        if (!encounterId.isNullOrBlank()) {
-          filter(
-            QuestionnaireResponse.ENCOUNTER,
-            {
-              value =
-                encounterId.extractLogicalIdUuid().asReference(ResourceType.Encounter).reference
-            },
-          )
-        }
-        if (!questionnaireResponseStatus.isNullOrBlank()) {
-          filter(
-            QuestionnaireResponse.STATUS,
-            { value = of(questionnaireResponseStatus) },
-          )
-        }
-      }
-    val questionnaireResponses: List<QuestionnaireResponse> = defaultRepository.search(search)
-    return questionnaireResponses.maxByOrNull { it.meta.lastUpdated }
-  }
-
   private suspend fun launchContextResources(
     subjectResourceType: ResourceType?,
     subjectResourceIdentifier: String?,
@@ -1176,7 +1135,8 @@ constructor(
             questionnaireConfig.isReadOnly() ||
             questionnaireConfig.saveDraft)
       ) {
-        searchQuestionnaireResponse(
+        defaultRepository
+          .searchQuestionnaireResponse(
             resourceId = resourceIdentifier,
             resourceType = resourceType,
             questionnaireId = questionnaire.logicalId,
